requirements/base.txt
asgiref==3.7.2
autopep8==2.0.4
Django==4.1.13
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.5.0
Markdown==3.5.2
phonenumbers==8.13.31
pillow==10.2.0
pycodestyle==2.11.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
tzdata==2023.4
